pyproject.toml
[project]
name = "audio-notes-app"
version = "0.1.0"
description = "Audio notes web application with FastAPI, PostgreSQL, RabbitMQ, and AI transcription"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.6.0",
    "python-multipart>=0.0.17",
    "aio-pika>=9.5.0",
    "httpx>=0.28.0",
    "supertokens-python>=0.28.0",
    "deepgram-sdk>=3.8.0",
    "google-generativeai>=0.8.3",
    "websockets>=15.0.1",
    "loguru>=0.7.0",
    "tenacity>=8.2.3",
    "slowapi>=0.1.9",
    "redis>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
test = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "fakeredis>=2.20.0",
]
dev = [
    "ruff>=0.1.9",
    "mypy>=1.7.1",
    "pre-commit>=3.6.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
